requirements.txt
allure-pytest==2.12.0
allure-python-commons==2.12.0
attrs==22.2.0
certifi==2022.12.7
cfgv==3.3.1
charset-normalizer==3.0.1
curlify==2.2.1
distlib==0.3.6
exceptiongroup==1.1.0
Faker==16.4.0
filelock==3.9.0
flake8==6.0.0
identify==2.5.13
idna==3.4
iniconfig==2.0.0
mccabe==0.7.0
nodeenv==1.7.0
packaging==23.0
platformdirs==2.6.2
pluggy==1.0.0
pre-commit==2.21.0
pycodestyle==2.10.0
pyflakes==3.0.1
pytest==7.2.0
pytest-voluptuous==1.2.0
python-dateutil==2.8.2
python-dotenv==0.21.0
PyYAML==6.0
requests==2.28.2
six==1.16.0
tomli==2.0.1
urllib3==1.26.14
virtualenv==20.17.1
voluptuous==0.13.1
